AutoStream's Chatbot/requirements.txt
langchain
langgraph
langchain-google-genai
google-generativeai
langchain-community
langchain-text-splitters
chromadb
python-dotenv
pydantic
langchain-huggingface
sentence-transformers
fastapi
uvicorn
